feat: remove moment, add nr1-community, closes newrelic/nr1-customer-journey#24
