+++ nerdlets/details/details.js
@@ -10,15 +10,13 @@ import {
   BlockText
 } from 'nr1';
 import { getJourneys } from '../../journeyConfig';
-import moment from 'moment';
-import momentDurationFormatSetup from 'moment-duration-format';
+import { timeRangeToNrql } from '@newrelic/nr1-community';
 
 const journeyConfig = getJourneys();
-momentDurationFormatSetup(moment);
 
 export default class Details extends React.Component {
   static propTypes = {
-    launcherUrlState: PropTypes.object.isRequired,
+    platformUrlState: PropTypes.object.isRequired,
     nerdletUrlState: PropTypes.object.isRequired
   };
 
@@ -28,13 +26,14 @@ export default class Details extends React.Component {
   }
 
   render() {
-    const { duration } = this.props.launcherUrlState.timeRange;
-    const durationInMinutes = duration / 60 / 1000;
+    const { platformUrlState, nerdletUrlState } = this.props;
+    const sinceStatement = timeRangeToNrql(platformUrlState);
+    const agoStatement = sinceStatement.split("SINCE")[1];
     const {
       selectedColumn,
       selectedJourney,
       selectedStep
-    } = this.props.nerdletUrlState;
+    } = nerdletUrlState;
     const journey = journeyConfig.find(j => j.id === selectedJourney);
     const column = journey.series.find(s => s.id === selectedColumn);
     const step = journey.steps.find(s => s.id === selectedStep);
@@ -52,8 +51,7 @@ export default class Details extends React.Component {
             step.nrqlWhere
           }) ${
             kpi.altNrql ? `AND (${kpi.altNrql}) ` : ''
-          } SINCE ${durationInMinutes} MINUTES AGO COMPARE WITH ${durationInMinutes *
-            2} MINUTES AGO`;
+          } ${sinceStatement} COMPARE WITH ${agoStatement}`;
           return kpi;
         });
     }
@@ -95,7 +93,7 @@ export default class Details extends React.Component {
               Timeseries Data
             </HeadingText>
             <BlockText className="timePeriodText">
-              Since {moment.duration(duration).format()} ago
+              {timeRangeToNrql(platformUrlState)}
             </BlockText>
           </GridItem>
           {stats.map((stat, i) => {
@@ -105,13 +103,11 @@ export default class Details extends React.Component {
                 step.altNrql.JavaScriptError
               }) AND (${
                 column.nrqlWhere
-              }) TIMESERIES SINCE ${durationInMinutes} MINUTES AGO COMPARE WITH ${durationInMinutes *
-                2} MINUTES AGO`;
+              }) TIMESERIES ${sinceStatement} COMPARE WITH ${agoStatement}`;
             } else {
               query = `${stat.value.nrql} AND (${step.nrqlWhere}) AND (${
                 column.nrqlWhere
-              }) TIMESERIES SINCE ${durationInMinutes} MINUTES AGO COMPARE WITH ${durationInMinutes *
-                2} MINUTES AGO`;
+              }) TIMESERIES ${sinceStatement} COMPARE WITH ${agoStatement}`;
             }
             // console.log(query);
             return (
