+++ package.json
@@ -17,9 +17,8 @@
     "sdkVersion": 2
   },
   "dependencies": {
+    "@newrelic/nr1-community": "^1.1.1",
     "lodash": "^4.17.15",
-    "moment": "^2.24.0",
-    "moment-duration-format": "^2.3.2",
     "nr1-funnel-component": "github:newrelic/nr1-funnel-component",
     "prop-types": "^15.6.2",
     "react": "^16.9.0",
